Implement CSS selector extraction and refactor research agent
- Added `cheerio` dependency for reliable HTML parsing.
- Created `worker/lib/html-utils.ts` with `extractBySelectors` and `extractFromHtml` utilities.
- Refactored `worker/lib/research-agent/fetcher.ts` (977 lines) into smaller modules: `api-fetchers.ts`, `generic-fetcher.ts`, `extractor.ts`, `transformers.ts`, and `rate-limiter.ts` to comply with the 500-line limit.
- Updated `ResearchSource` type to support CSS selectors.
- Integrated selector-based extraction into `extractReferralsFromContent` (research agent) and `parseHTMLContent` (discovery pipeline).
- Added unit tests in `tests/unit/html-utils.test.ts`.
- Fixed TypeScript and formatting issues.
- Verified system stability with existing integration tests.
