+++ package.json
@@ -32,6 +32,7 @@
     "verify": "bash scripts/quality_gate.sh"
   },
   "dependencies": {
+    "cheerio": "^1.2.0",
     "protobufjs": "8.3.0",
     "zod": "^3.22.4"
   },
